requirements.txt
# Core dependencies
faster-whisper>=1.0.3,<2.0
diart>=0.9.0,<1.1
onnxruntime>=1.17.0,<2.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# GUI
tkinter-tooltip>=2.1.0

# Optional for ChatGPT postprocessing  
openai>=1.0,<2.0

# Configuration and logging
pyyaml>=6.0
loguru>=0.7.0

# Development and testing
pytest>=7.0,<9.0
black>=23.0
isort>=5.12
flake8>=6.0
